refactor: extract FilterOptions component to new file
