+++ src/views/components/FilterOptions/IFilterOptions.tsx
@@ -1,0 +1,7 @@
+import { Genre, Platform, Publisher, Tag } from '@data/types'
+
+
+export interface IFilterOptions {
+	options: Publisher[] | Genre[] | Tag[] | Platform[]
+	typeName: string
+}
